Toxo to main (#79)
* major commit for vacuole functions

* fix bounds

* weird mistakes in eval segmentation, to review

* merge with main

* final vacuole cleanup

* remove all toxo specific code

---------
